// aah_apr_front_end-master/types/Types.ts
export type KitFormType = {
  kitId: string;
  destribution_date: string;
  remark: string;
  is_received: boolean;
};